Add functional testing
Add test bundles for 18.04 + Queens, Stein and Train releases of
OpenStack.

This change also includes features to ensure that the charm remains
in blocked status until:

  nfs-shares configuration has been supplied
  the create-cloud-admin-trust action has been executed
  the create-license action has been executed

This ensures that the charm can be automatically tested and users get
accurate feedback on the current status of the workload during
deployment and configuration.

Class structure and release determination has been shuffled about
in the trilio-* charms to support this set of changes.

Func-Test-Pr: https://github.com/openstack-charmers/zaza-openstack-tests/pull/252
Change-Id: I8abedb1ec253be07810580da2231721ae006e267